# src/main/resources/布灵4.0.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "布灵",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "用户登录",
        "deprecated": false,
        "description": "支持账号密码和手机号验证码两种登录方式. 成功后返回JWT令牌和用户信息.",
        "operationId": "loginUser",
        "tags": [
          "认证 (Authentication)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLoginResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "认证失败, 用户名或密码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/auth/register": {
      "post": {
        "summary": "用户注册",
        "deprecated": false,
        "description": "支持用户名密码和手机号验证码两种注册方式.",
        "operationId": "registerUser",
        "tags": [
          "认证 (Authentication)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "注册失败, 如用户已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/users/{userId}/profile": {
      "get": {
        "summary": "查看指定用户主页",
        "deprecated": false,
        "description": "需额外展示当前自己是否已关注该用户从而显示关注,取关按钮",
        "tags": [],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "状态码 (200:成功, 其他:失败)",
                      "examples": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息",
                      "examples": [
                        "Success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "avatarUrl": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "location": {
                          "type": "string"
                        },
                        "followerCount": {
                          "type": "integer"
                        },
                        "followedCount": {
                          "type": "integer"
                        },
                        "isFollowed": {
                          "type": "boolean",
                          "description": "当前登录用户是否关注了被访问用户"
                        }
                      },
                      "required": [
                        "id",
                        "nickname"
                      ],
                      "nullable": true
                    },
                    "knowledgeBases": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "int64"
                          },
                          "iconIndex": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "shareId": {
                            "type": "string",
                            "nullable": true
                          },
                          "coverUrl": {
                            "type": "string",
                            "format": "uri",
                            "nullable": true
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "knowledgeBases"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/profile": {
      "get": {
        "summary": "查看当前登录用户的个人信息",
        "deprecated": false,
        "description": "",
        "operationId": "getCurrentUserProfile",
        "tags": [
          "用户 (User)"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserProfile"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新当前登录用户的个人信息",
        "deprecated": false,
        "description": "",
        "operationId": "updateCurrentUserProfile",
        "tags": [
          "用户 (User)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdate"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "状态码 (200:成功, 其他:失败)",
                      "examples": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息",
                      "examples": [
                        "Success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "int64"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "avatarUrl": {
                          "type": "string",
                          "format": "uri"
                        },
                        "discription": {
                          "type": "string"
                        },
                        "location": {
                          "type": "string"
                        },
                        "field": {
                          "type": "string"
                        },
                        "followerCount": {
                          "type": "integer"
                        },
                        "followedCount": {
                          "type": "integer"
                        }
                      },
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/{userId}/public-kbs": {
      "get": {
        "summary": "查看指定用户的公开知识库展示",
        "deprecated": false,
        "description": "用于个人主页, 展示最多三个公开的知识库.",
        "operationId": "getUserPublicKbs",
        "tags": [
          "用户 (User)"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPublicKbList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/user/history/views": {
      "get": {
        "summary": "获取浏览历史记录",
        "deprecated": false,
        "description": "获取当前用户的文档浏览历史记录，按浏览时间倒序排列。",
        "operationId": "getBrowsingHistory",
        "tags": [
          "用户历史 (User History)"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultActivityRecordList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "新增浏览记录",
        "deprecated": true,
        "description": "记录一次文档浏览事件。推荐在前端打开文档详情页时显式调用此接口。",
        "operationId": "addBrowsingRecord",
        "tags": [
          "用户历史 (User History)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewHistoryCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "记录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/history/likes": {
      "get": {
        "summary": "获取点赞历史记录",
        "deprecated": false,
        "description": "获取当前用户点赞过的所有文档列表，按点赞时间倒序排列。",
        "operationId": "getLikeHistory",
        "tags": [
          "用户历史 (User History)"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultActivityRecordList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/history/comments": {
      "get": {
        "summary": "获取评论历史记录",
        "deprecated": false,
        "description": "获取当前用户评论过的所有文档列表，按最后评论时间倒序排列。",
        "operationId": "getCommentHistory",
        "tags": [
          "用户历史 (User History)"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultActivityRecordList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/users/{userId}/follow": {
      "post": {
        "summary": "关注用户",
        "deprecated": false,
        "description": "",
        "operationId": "followUser",
        "tags": [
          "用户社交 (User Social)"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "被关注者的用户ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "取消关注用户",
        "deprecated": false,
        "description": "",
        "operationId": "unfollowUser",
        "tags": [
          "用户社交 (User Social)"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "被取消关注者的用户ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/users/{userId}/following": {
      "get": {
        "summary": "获取用户关注的列表 (Following)",
        "deprecated": false,
        "description": "关注列表肯定默认关注了,不必关心isfollowing",
        "operationId": "listFollowingUsers",
        "tags": [
          "用户社交 (User Social)"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "要查询的用户ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserSummaryList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/users/{userId}/followers": {
      "get": {
        "summary": "获取用户的粉丝列表 (Followers)",
        "deprecated": false,
        "description": "",
        "operationId": "listFollowerUsers",
        "tags": [
          "用户社交 (User Social)"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "要查询的用户ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserSummaryList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200:成功, 其他:失败)",
            "examples": [
              200
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "examples": [
              "Success"
            ]
          },
          "data": {
            "type": "object",
            "description": "响应数据",
            "properties": {},
            "nullable": true
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "loginType",
          "identifier",
          "credential"
        ],
        "properties": {
          "loginType": {
            "type": "integer",
            "description": "登录类型: 1-账号密码, 2-手机号+验证码",
            "examples": [
              1
            ]
          },
          "identifier": {
            "type": "string",
            "description": "登录标识 (用户名 或 手机号)",
            "examples": [
              "testuser"
            ]
          },
          "credential": {
            "type": "string",
            "description": "凭证 (密码 或 短信验证码)",
            "examples": [
              "password123"
            ]
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "registerType",
          "identifier",
          "password"
        ],
        "properties": {
          "registerType": {
            "type": "integer",
            "description": "注册类型: 1-用户名, 2-手机号",
            "examples": [
              1
            ]
          },
          "identifier": {
            "type": "string",
            "description": "注册标识 (用户名 或 手机号)",
            "examples": [
              "newuser"
            ]
          },
          "password": {
            "type": "string",
            "description": "密码",
            "examples": [
              "password123"
            ]
          },
          "verificationCode": {
            "type": "string",
            "description": "手机注册时的验证码",
            "examples": [
              "123456"
            ]
          }
        }
      },
      "UserProfileUpdate": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "examples": [
              "新的昵称"
            ]
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "examples": [
              "http://example.com/new_avatar.png"
            ]
          },
          "discription": {
            "type": "string",
            "examples": [
              "这是一段新的简介"
            ]
          },
          "location": {
            "type": "string",
            "examples": [
              "北京"
            ]
          },
          "field": {
            "type": "string",
            "examples": [
              "互联网"
            ]
          }
        }
      },
...
      "LoginResponseData": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            ]
          },
          "username": {
            "type": "string",
            "examples": [
              "testuser"
            ]
          },
          "nickname": {
            "type": "string",
            "examples": [
              "测试用户"
            ]
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "examples": [
              "http://example.com/avatar.png"
            ]
          }
        }
      },
      "UserProfileData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "int64"
          },
          "nickname": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          },
          "discription": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "followerCount": {
            "type": "integer"
          },
          "followedCount": {
            "type": "integer"
          }
        }
      },
      "PublicKnowledgeBase": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "examples": [
              "公开的Go语言学习笔记"
            ]
          },
          "description": {
            "type": "string",
            "examples": [
              "这里记录了Go语言学习过程中的笔记"
            ]
          },
          "coverUrl": {
            "type": "string",
            "format": "uri"
          },
          "viewCount": {
            "type": "integer"
          }
        }
      },
...
      "ResultLoginResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/LoginResponseData"
              }
            }
          }
        ]
      },
      "ResultUserProfile": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UserProfileData"
              }
            }
          }
        ]
      },
      "ResultPublicKbList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PublicKnowledgeBase"
                }
              }
            }
          }
        ]
      },
...
      "ResultActivityRecordList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ActivityRecord"
                }
              }
            }
          }
        ]
      },
      "ActivityRecord": {
        "type": "object",
        "description": "通用的历史记录对象结构",
        "properties": {
          "resourceId": {
            "type": "string",
            "description": "文档的ID"
          },
          "resourceTitle": {
            "type": "string",
            "description": "文档的标题"
          },
          "resourceType": {
            "type": "string",
            "description": "文档的类型 (e.g., markdown)"
          },
          "kbId": {
            "type": "string",
            "description": "所属知识库的ID"
          },
          "kbName": {
            "type": "string",
            "description": "所属知识库的名称"
          },
          "authorId": {
            "type": "string",
            "description": "文档作者的用户ID"
          },
          "authorName": {
            "type": "string",
            "description": "文档作者的昵称"
          },
          "actionAt": {
            "type": "string",
            "format": "date-time",
            "description": "操作发生的时间 (UTC)"
          }
        }
      },
      "ViewHistoryCreateRequest": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "description": "被浏览的文档ID"
          }
        }
      },
...
      "ResultUserSummaryList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}